Move template resolution to a classpath loader, add routes to view other peoples profiles, update template to render both profile versions
