+++ src/main/java/space/pxls/server/UndertowServer.java
@@ -89,13 +89,12 @@ public class UndertowServer {
                 .addPrefixPath("/notifications", webHandler::notificationsList)
                 .addExactPath("/", webHandler::index)
                 .addExactPath("/index.html", webHandler::index)
-                .addExactPath("/profile", Handlers.redirect("/profile/view"))
-                .addExactPath("/profile/index.html", Handlers.redirect("/profile/view"))
-                .addExactPath("/profile/view", webHandler::profile)
                 .addExactPath("/factions", new AllowedMethodsHandler(webHandler::getRequestingUserFactions, Methods.GET))
                 .addPrefixPath("/", Handlers.resource(new ClassPathResourceManager(App.class.getClassLoader(), "public/"))
                         .setCacheTime(10));
         RoutingHandler routingHandler = Handlers.routing()
+            .get("/profile", webHandler::profileView)
+            .get("/profile/{who}", webHandler::profileView)
             .get("/factions/{fid}", new JsonReader(webHandler::manageFactions))
             .post("/factions", new JsonReader(webHandler::manageFactions))
             .put("/factions/{fid}", new JsonReader(webHandler::manageFactions))
